+++ src/Server/Persistence/Database/Sqlite/Queries/SqliteRetrieveEntityQuery.cs
@@ -32,23 +32,23 @@ public sealed class SqliteRetrieveEntityQuery : IRetrieveEntityQuery
     private const string query =
         @"WITH RECURSIVE EntityTree(id, template_id, x, y, z, material, materialModifier, playerCharacter, name, account, 
                 parent, drawable, animatedSprite, orientation, admin, castBlockShadows, plsRadius, plsIntensity,
-                plsColorR, plsColorG, plsColorB, plsPosX, plsPosY, plsPosZ)
+                plsColor, plsPosX, plsPosY, plsPosZ)
 	        AS (
 		        SELECT id, template_id, x, y, z, material, materialModifier, playerCharacter, name, account, parent, 
                         drawable, animatedSprite, orientation, admin, castBlockShadows, plsRadius, plsIntensity,
-                        plsColorR, plsColorG, plsColorB, plsPosX, plsPosY, plsPosZ
+                        plsColor, plsPosX, plsPosY, plsPosZ
 			        FROM EntityWithComponents WHERE id = @Id
 		        UNION ALL
             		SELECT ec.id, ec.template_id, ec.x, ec.y, ec.z, ec.material, ec.materialModifier, ec.playerCharacter,
                         ec.name, ec.account, ec.parent, ec.drawable, ec.animatedSprite, ec.orientation, ec.admin,
-                        ec.castBlockShadows, ec.plsRadius, ec.plsIntensity, ec.plsColorR, ec.plsColorG, ec.plsColorB,
+                        ec.castBlockShadows, ec.plsRadius, ec.plsIntensity, ec.plsColor,
                         ec.plsPosX, ec.plsPosY, ec.plsPosZ
 			        FROM EntityWithComponents ec, EntityTree et
         			WHERE ec.parent = et.id
 	        )
             SELECT id, template_id, x, y, z, material, materialModifier, playerCharacter, name, account, parent, drawable,
-                animatedSprite, orientation, admin, castBlockShadows, plsRadius, plsIntensity, plsColorR, plsColorG,
-                plsColorB, plsPosX, plsPosY, plsPosZ
+                animatedSprite, orientation, admin, castBlockShadows, plsRadius, plsIntensity, plsColor,
+                plsPosX, plsPosY, plsPosZ
             FROM EntityTree ORDER BY parent NULLS LAST";
 
     private readonly SqliteConnection dbConnection;